In ISearchable interface was changed the type of list count from int to long

+++ src/Eclo.DataAccess/Repositories/Products/ProductRepository.cs
@@ -115,7 +115,7 @@ public class ProductRepository : BaseRepository, IProductRepository
         }
     }
 
-    public async Task<(int ItemsCount, IList<ProductViewModel>)> SearchAsync(string search, PaginationParams @params)
+    public async Task<(long ItemsCount, IList<ProductViewModel>)> SearchAsync(string search, PaginationParams @params)
     {
         try
         {
